feat: FoodEntriesController

+++ FoodTracker.Application/Features/Activities/CreateActivityItemCommand.cs
@@ -8,7 +8,7 @@ public sealed class CreateActivityTypeCommand : IRequest<Result<ActivityTypeDto>
 {
     public string Name { get; init; } = string.Empty;
     public string? Description { get; init; }
-    public decimal? CaloriesPerHour { get; init; }
-    public decimal? CaloriesPer100Reps { get; init; }
+    public double? CaloriesPerHour { get; init; }
+    public double? CaloriesPer100Reps { get; init; }
     public string? Category { get; init; }
 }
